+++ variables.tf
@@ -154,7 +154,7 @@ variable "cluster_delete_timeout" {
 }
 
 variable "local_exec_interpreter" {
-  description = "Command to run for local-exec resources. Must be a shell-style interpreter. If you are on Windows Git Bash is a good choice. Defaults to [\"/bin/sh\", \"-c\"]"
+  description = "Command to run for local-exec resources. Must be a shell-style interpreter. If you are on Windows Git Bash is a good choice."
   type        = "list"
   default     = ["/bin/sh", "-c"]
 }