+++ aws_auth.tf
@@ -42,7 +42,7 @@ data "template_file" "config_map_aws_auth" {
 }
 
 data "template_file" "map_users" {
-  count    = "${length(var.map_users)}"
+  count    = "${var.map_users_count}"
   template = "${file("${path.module}/templates/config-map-aws-auth-map_users.yaml.tpl")}"
 
   vars {
@@ -53,7 +53,7 @@ data "template_file" "map_users" {
 }
 
 data "template_file" "map_roles" {
-  count    = "${length(var.map_roles)}"
+  count    = "${var.map_roles_count}"
   template = "${file("${path.module}/templates/config-map-aws-auth-map_roles.yaml.tpl")}"
 
   vars {
@@ -64,7 +64,7 @@ data "template_file" "map_roles" {
 }
 
 data "template_file" "map_accounts" {
-  count    = "${length(var.map_accounts)}"
+  count    = "${var.map_accounts_count}"
   template = "${file("${path.module}/templates/config-map-aws-auth-map_accounts.yaml.tpl")}"
 
   vars {