+++ examples/eks_test_fixture/variables.tf
@@ -12,6 +12,12 @@ variable "map_accounts" {
   ]
 }
 
+variable "map_accounts_count" {
+  description = "The count of accounts in the map_accounts list."
+  type        = "string"
+  default     = 2
+}
+
 variable "map_roles" {
   description = "Additional IAM roles to add to the aws-auth configmap."
   type        = "list"
@@ -23,6 +29,12 @@ variable "map_roles" {
       group    = "system:masters"
     },
   ]
+}
+
+variable "map_roles_count" {
+  description = "The count of roles in the map_roles list."
+  type        = "string"
+  default     = 1
 }
 
 variable "map_users" {
@@ -41,4 +53,10 @@ variable "map_users" {
       group    = "system:masters"
     },
   ]
+}
+
+variable "map_users_count" {
+  description = "The count of roles in the map_users list."
+  type        = "string"
+  default     = 2
 }